refactor: mock data generator
